fix(medical): correct magic_unpair precision and overlap detection
- Change magic_unpair to use f64 instead of f32 for sqrt, preventing
  precision loss with SNOMED-scale IDs (100M-900M range)
- Fix incomplete overlap detection in umls_extractor and medical_extractor
  that missed the containment case (new match fully containing existing)

+++ crates/terraphim_rolegraph/src/lib.rs
@@ -946,7 +946,7 @@ pub fn magic_pair(x: u64, y: u64) -> u64 {
 /// }
 #[memoize(CustomHasher: ahash::AHashMap)]
 pub fn magic_unpair(z: u64) -> (u64, u64) {
-    let q = (z as f32).sqrt().floor() as u64;
+    let q = (z as f64).sqrt().floor() as u64;
     let l = z - q * q;
     if l < q { (l, q) } else { (q, l - q) }
 }
